requirements.txt
fire>=0.4.0
requests>=2.27.1
pyfiglet>=0.8.post1
progressbar2>=4.0.0
python-slugify>=5.0.2
beautifulsoup4>=4.10.0